# Improve performance of RPKUM normalization in featureCounts