--- a/resolwe_bio/processes/expression/featureCounts.yml
+++ b/resolwe_bio/processes/expression/featureCounts.yml
@@ -14,7 +14,7 @@
       cores: 10
       network: true
   data_name: "{{ alignment.aligned_reads|sample_name|default('?') }}"
-  version: 2.4.1
+  version: 2.4.2
   type: data:expression:featurecounts
   category: Quantify
   flow_collection: sample