Transition from `pandas.DataFrame` to `pandas.Series`.