@mzganec 2019-07-15
Benchmark was performed using expression file containing 12,869 genes and mappability file with 12,883 genes. The following run times were obtained for the computational section of the code, excluding reading input files and writing output file.

| old code | new code |
|:--------:|:--------:|
| 1046 ms  | 53 ms  |