@jkokosar 2019-07-15
Considerable improvement! Thanks!